Abstract components from Home and House
