+++ src/components/atoms.tsx
@@ -1,6 +1,8 @@
 import { mq, Typography } from '@ensdomains/thorin';
 import styled, { css, DefaultTheme } from 'styled-components';
 
+import Anchor from './Anchor';
+
 export const cardStyles = ({ theme, hasPadding }: { theme: DefaultTheme; hasPadding?: boolean }) => css`
   display: flex;
   flex-direction: column;
@@ -52,5 +54,117 @@ export const DisplayItems = styled.div(
     justify-content: stretch;
     width: ${theme.space.full};
     gap: ${theme.space['2']};
+  `
+);
+
+export const SectionHeading = styled.div(
+  ({ theme }) => css`
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: center;
+    padding: ${theme.space['2']};
+
+    ${mq.md.min(css`
+      flex-direction: row;
+      align-items: center;
+      justify-content: space-between;
+      margin-bottom: ${theme.space['2']};
+      gap: ${theme.space['4']};
+
+      padding: 0;
+    `)}
+
+    width: 100%;
+
+    &.desktop-only {
+      ${mq.md.max(css`
+        display: none;
+      `)}
+    }
+  `
+);
+
+export const ActiveTypography = styled(Typography)(
+  ({ theme }) => css`
+    color: ${theme.colors.textTertiary};
+    font-weight: bold;
+  `
+);
+
+export const MobileHiddenAnchor = styled(Anchor)(
+  () => css`
+    display: none;
+    ${mq.md.min(css`
+      display: block;
+    `)}
+  `
+);
+
+export const DesktopHiddenAnchor = styled(Anchor)(
+  () => css`
+    display: block;
+    ${mq.md.min(css`
+      display: none;
+    `)}
+  `
+);
+
+export const Heading = styled(Typography)(
+  ({ theme }) => css`
+    font-size: ${theme.space['9']};
+    font-weight: bold;
+  `
+);
+
+export const Subheading = styled(Typography)(
+  ({ theme }) => css`
+    color: ${theme.colors.textTertiary};
+    font-size: ${theme.fontSizes.large};
+  `
+);
+
+export const HeadingContainer = styled.div(
+  ({ theme }) => css`
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+
+    gap: ${theme.space['2']};
+    max-width: ${theme.space['144']};
+    text-align: center;
+
+    margin-top: ${theme.space['10']};
+
+    ${mq.md.min(css`
+      margin-top: 0;
+    `)}
+  `
+);
+
+export const RoundGrid = styled.div(
+  ({ theme }) => css`
+    display: grid;
+    grid-template-columns: repeat(1, 1fr);
+    gap: ${theme.space['8']};
+    width: 100%;
+
+    ${mq.md.min(css`
+      grid-template-columns: repeat(2, 1fr);
+    `)}
+  `
+);
+
+export const RoundItemsOuter = styled.div(
+  ({ theme }) => css`
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: ${theme.space['2']};
+    width: 100%;
+    max-width: ${theme.space['256']};
+    margin-top: ${theme.space['2']};
   `
 );
